Corrigido bug na animação, corrigido bug em que o jogador continuaria se movimentando se fosse apertado a tecla alt enquanto o jogador estivesse apertando a tecla de movimento. Essa correção do bug do movimento é tapa buraco, o certo é fazer uma tela de pausa e pausar o jogo se o cara apertar alt

+++ src/SpaceInvadersClone/Sprite.java
@@ -7,7 +7,7 @@ public class Sprite extends SpriteSheet {
     }
 
     @Override
-    public void animacao(int periodo, double tempoTotal) {
+    public void animacao(double periodo, double tempoTotal) {
         if (tempoTotal == periodo) {
             if (this.frame < this.sprites.length - 1) {
                 this.frame++;
